+++ R/track-2d.R
@@ -82,6 +82,8 @@ gtrack.2d.create <- function(track = NULL, description = NULL, intervals = NULL,
     }
     .gcheckroot()
 
+    intervals <- rescue_ALLGENOME(intervals, as.character(substitute(intervals)))
+
     trackstr <- do.call(.gexpr2str, list(substitute(track)), envir = parent.frame())
     intervalsstr <- deparse(substitute(intervals), width.cutoff = 500)[1]
     valuesstr <- deparse(substitute(values), width.cutoff = 500)[1]